NonStopBridge/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.6
pyModbusTCP==0.2.1
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.1.0
